# Optimize color matching