--- a/src/ui/ccModal.ts
+++ b/src/ui/ccModal.ts
@@ -1,7 +1,8 @@
 /// <reference types="tampermonkey" />
 import { WPLACE_FREE, WPLACE_PAID, WPLACE_NAMES, DEFAULT_FREE_KEYS } from '../core/palette';
 import { createCanvas } from '../core/canvas';
-import { config, saveConfig } from '../core/store';
+import { colorCaches } from '../core/cache';
+import { config, saveConfig, type OverlayItem } from '../core/store';
 import { MAX_OVERLAY_DIM } from '../core/constants';
 import { ensureHook } from '../core/hook';
 import { clearOverlayCache, paletteDetectionCache } from '../core/cache';
@@ -38,7 +39,7 @@ type CCState = {
   selectedPaid: Set<string>;
   realtime: boolean;
 
-  overlay: any | null;
+  overlay: OverlayItem | null;
   lastColorCounts: Record<string, number>;
   isStale: boolean;
 };
@@ -191,7 +192,7 @@ export function buildCCModal() {
   renderPaletteGrid();
 }
 
-export function openCCModal(overlay: any) {
+export function openCCModal(overlay: OverlayItem) {
   if (!cc) return;
   cc.overlay = overlay;
 
@@ -266,7 +267,12 @@ function processImage() {
 
   const palette = getActivePalette();
   const counts: Record<string, number> = {};
-  const colorCache: Map<number, [number, number, number]> = new Map();
+  const id = cc.overlay.id;
+  let colorCache = colorCaches.get(id);
+  if (!colorCache) {
+    colorCache = new Map();
+    colorCaches.set(id, colorCache);
+  }
 
   for (let i = 0; i < src.length; i += 4) {
     const r = src[i], g = src[i+1], b = src[i+2], a = src[i+3];