This pr introdus 2 optimizations for color matching
- remove unnecssary `Math.sqrt` for color distance comaprision(this is a cpu-bound operation, reduces from ~18s to ~3s on a over 1000 x 1000 image)
- cache color calculation on the assumption that there are many same pixel(reduces from ~3s to ~1s on a over 1000 x 1000 image with the 1st optimization).